+++ jib-core/src/main/java/com/google/cloud/tools/jib/builder/BuildImageSteps.java
@@ -23,6 +23,7 @@ import com.google.cloud.tools.jib.cache.CacheMetadataCorruptedException;
 import com.google.cloud.tools.jib.cache.CachedLayer;
 import com.google.cloud.tools.jib.http.Authorization;
 import com.google.cloud.tools.jib.image.Image;
+import com.google.cloud.tools.jib.registry.credentials.RegistryCredentials;
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.util.concurrent.Futures;
 import com.google.common.util.concurrent.ListenableFuture;
@@ -66,10 +67,29 @@ public class BuildImageSteps {
             MoreExecutors.listeningDecorator(Executors.newCachedThreadPool());
 
         try (Cache cache = Cache.init(cacheDirectory)) {
+          timer2.lap("Setting up credential retrieval");
+          ListenableFuture<RegistryCredentials> retrieveRegistryCredentialsFuture =
+              listeningExecutorService.submit(
+                  new RetrieveRegistryCredentialsStep(buildConfiguration));
+
+          timer2.lap("Setting up image push authentication");
+          // Authenticates push.
+          ListenableFuture<Authorization> authenticatePushFuture =
+              Futures.whenAllSucceed(retrieveRegistryCredentialsFuture)
+                  .call(
+                      new AuthenticatePushStep(
+                          buildConfiguration, retrieveRegistryCredentialsFuture),
+                      listeningExecutorService);
+
           timer2.lap("Setting up image pull authentication");
           // Authenticates base image pull.
           ListenableFuture<Authorization> authenticatePullFuture =
-              listeningExecutorService.submit(new AuthenticatePullStep(buildConfiguration));
+              Futures.whenAllSucceed(retrieveRegistryCredentialsFuture)
+                  .call(
+                      new AuthenticatePullStep(
+                          buildConfiguration, retrieveRegistryCredentialsFuture),
+                      listeningExecutorService);
+
           timer2.lap("Setting up base image pull");
           // Pulls the base image.
           ListenableFuture<Image> pullBaseImageFuture =
@@ -90,10 +110,6 @@ public class BuildImageSteps {
                           pullBaseImageFuture),
                       listeningExecutorService);
 
-          timer2.lap("Setting up image push authentication");
-          // Authenticates push.
-          ListenableFuture<Authorization> authenticatePushFuture =
-              listeningExecutorService.submit(new AuthenticatePushStep(buildConfiguration));
           timer2.lap("Setting up base image layer push");
           // Pushes the base image layers.
           ListenableFuture<List<ListenableFuture<Void>>> pushBaseImageLayerFuturesFuture =