+++ jib-core/src/main/java/com/google/cloud/tools/jib/registry/RegistryAuthenticator.java
@@ -31,6 +31,7 @@ import java.net.MalformedURLException;
 import java.net.URL;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
+import javax.annotation.Nullable;
 
 /**
  * Authenticates pull access with a registry service.
@@ -39,8 +40,6 @@ import java.util.regex.Pattern;
  *     href="https://docs.docker.com/registry/spec/auth/token/">https://docs.docker.com/registry/spec/auth/token/</a>
  */
 public class RegistryAuthenticator {
-
-  private final URL authenticationUrl;
 
   // TODO: Replace with a WWW-Authenticate header parser.
   /**
@@ -92,21 +91,35 @@ public class RegistryAuthenticator {
     private String token;
   }
 
+  private final URL authenticationUrl;
+  @Nullable private Authorization authorization;
+
   RegistryAuthenticator(String realm, String service, String repository)
       throws MalformedURLException {
     authenticationUrl =
         new URL(realm + "?service=" + service + "&scope=repository:" + repository + ":pull");
   }
 
+  /** Sets an {@code Authorization} header to authenticate with. */
+  public RegistryAuthenticator setAuthorization(@Nullable Authorization authorization) {
+    this.authorization = authorization;
+    return this;
+  }
+
   /** Sends the authentication request and retrieves the Bearer authorization token. */
   public Authorization authenticate() throws RegistryAuthenticationFailedException {
     try (Connection connection = new Connection(authenticationUrl)) {
-      Response response = connection.get(Request.builder().build());
+      Request.Builder requestBuilder = Request.builder();
+      if (authorization != null) {
+        requestBuilder.setAuthorization(authorization);
+      }
+      Response response = connection.get(requestBuilder.build());
       String responseString = Blobs.writeToString(response.getBody());
 
       AuthenticationResponseTemplate responseJson =
           JsonTemplateMapper.readJson(responseString, AuthenticationResponseTemplate.class);
       return Authorizations.withBearerToken(responseJson.token);
+
     } catch (IOException ex) {
       throw new RegistryAuthenticationFailedException(ex);
     }