+++ CHANGELOG.md
@@ -14,7 +14,12 @@ All notable changes to this project will be documented in this file.
 - Better error messages when pushing an image manifest ([#63](https://github.com/google/jib/pull/63))
 - Validates target image configuration ([#63](https://github.com/google/jib/pull/63))
 
+### Changed
+- Configure multiple credential helpers with `credHelpers` ([#68](https://github.com/google/jib/pull/68))
+- Removed configuration `credentialHelperName` ([#68](https://github.com/google/jib/pull/68))
+
 ### Fixed
 - Build failure on Windows ([#74](https://github.com/google/jib/issues/74))
 - Infers common credential helper names (for GCR and ECR) ([#64](https://github.com/google/jib/pull/64))
+- Cannot use private base image ([#68](https://github.com/google/jib/pull/68))
 - Building applications with no resources ([#73](https://github.com/google/jib/pull/73))